model average forecast
Added forecast year's data to prep data output. Then moved on to calculating a model average forecast as part of the model selection criteria. Trying to simplify the hindcast-forecast process but in this version I still need to create separate forecasts for each method. This seems redundant. Going to try doing hindcasts and forecasts at the same time. It that does not work I will revert to this commit.